separate auto advance section
