+++ src/app/posts/[slug]/page.tsx
@@ -4,10 +4,10 @@ import LikeButtonItem from '@/components/post/likeButtonItem';
 import PostContentSection from '@/components/post/postContentSection';
 import PostDispatcher from '@/components/post/postDispatcher';
 import PostHeaderSection from '@/components/post/postHeaderSection';
-import PostViewer from '@/components/post/postViewer';
-import PostViewerModeButton from '@/components/post/postViewerModeButton';
 import PostViewTracker from '@/components/post/postViewTracker';
 import TableOfContentsItem from '@/components/post/tableOfContentsItem';
+import PostViewer from '@/components/postViewer/postViewer';
+import PostViewerModeButton from '@/components/postViewer/postViewerModeButton';
 import { Post } from '@/features/post/domain/model/post';
 import { fetchPostBySlug } from '@/features/post/domain/service/postService';
 import { Suspense } from 'react';
